address feedback from code review
