+++ packages/wonder-blocks-core/index.js
@@ -1,10 +1,9 @@
 // @flow
-import ClickableBehavior, {
-    getClickableBehavior,
-} from "./components/clickable-behavior.js";
+import ClickableBehavior from "./components/clickable-behavior.js";
 import Text from "./components/text.js";
 import View from "./components/view.js";
 import addStyle from "./util/add-style.js";
+import getClickableBehavior from "./util/get-clickable-behavior.js";
 import type {ClickableHandlers} from "./components/clickable-behavior.js";
 import type {AriaProps, TextTag} from "./util/types.js";
